Validation done on Add and Delete buttons

+++ src/com/hscube/FrameMain.java
@@ -145,8 +145,10 @@ public class FrameMain extends JFrame {
 	private void btnDeleteWishListActionPerformed(ActionEvent e) {
 		ArrayList<Integer> arrList = new ArrayList<Integer>();
 		ArrayList<Integer> arrListRemove = new ArrayList<Integer>();
+		boolean hasRow = false;
 
 		for (int i = 0; i < tblWishlist.getRowCount(); i++) {
+			hasRow = true;
 			Boolean chked = Boolean
 					.valueOf(tblWishlist.getValueAt(i, 1).toString());
 			int ID = (int) tblWishlist.getValueAt(i, 0);
@@ -157,7 +159,17 @@ public class FrameMain extends JFrame {
 				tblWishlist.setValueAt(Boolean.FALSE, i, 1);
 			}
 		}
+		
+		if(!hasRow) {
+			JOptionPane.showMessageDialog(this, "No item to delete.");
+			return;
+		}
 
+		if(arrListRemove.size() == 0) {
+			JOptionPane.showMessageDialog(this, "Item not selected.");
+			return;
+		}
+		
 		if (WishlistController.deleteItemFromWishList(arrList)) {
 			insertTableRowsInWishList((DefaultTableModel) tblWishlist.getModel());
 			
@@ -168,10 +180,13 @@ public class FrameMain extends JFrame {
 	
 
 	private void btnDeleteActionPerformed(ActionEvent e) {
+		boolean hasRow = false;
+		
 		ArrayList<Integer> arrList = new ArrayList<Integer>();
 		ArrayList<Integer> arrListRemove = new ArrayList<Integer>();
 
 		for (int i = 0; i < tblItem.getRowCount(); i++) {
+			hasRow = true;
 			Boolean chked = Boolean
 					.valueOf(tblItem.getValueAt(i, 1).toString());
 			int ID = (int) tblItem.getValueAt(i, 0);
@@ -182,6 +197,17 @@ public class FrameMain extends JFrame {
 				tblItem.setValueAt(Boolean.FALSE, i, 1);
 			}
 		}
+		
+		if(!hasRow) {
+			JOptionPane.showMessageDialog(this, "No item to delete.");
+			return;
+		}
+		
+		if(arrListRemove.size() == 0) {
+			JOptionPane.showMessageDialog(this, "Item not selected.");
+			return;
+		}
+		
 
 		if (ItemController.deleteItems(arrList)) {
 			insertTableRows((DefaultTableModel) tblItem.getModel());
@@ -654,6 +680,11 @@ public class FrameMain extends JFrame {
 
 		private void btnAddActionPerformed(ActionEvent e) {
 			// add the item to the database
+			
+			if(txtName.getText().trim().equals("")) {
+				JOptionPane.showMessageDialog(this, "Enter the name for an item.");
+				return;
+			}
 
 			Item item = new Item(0, txtName.getText(),
 					CategoryController.getCategoryByName(comboboxCategory
